수업 상세 데이터 연결
